Rename navigationConfirmation to confirmation
