+++ packages/root/types/confirmation.d.ts
@@ -1,0 +1,2 @@
+import { ConfirmationMethods } from "./types/confirmation";
+export default function confirmation(): ConfirmationMethods;
