+++ packages/root/types/index.d.ts
@@ -2,6 +2,7 @@ export * from "./types";
 import locationUtils from "./locationUtils";
 import keyGenerator from "./keyGenerator";
 import navigateWith from "./navigateWith";
-import navigationConfirmation from "./navigationConfirmation";
+import confirmation from "./confirmation";
 import createBase from "./base";
-export { locationUtils, keyGenerator, navigationConfirmation, navigateWith, createBase };
+declare const navigationConfirmation: typeof confirmation;
+export { locationUtils, keyGenerator, confirmation, navigationConfirmation, navigateWith, createBase };
