--- packages/root/types/navigationConfirmation.d.ts
@@ -1,2 +1,0 @@
-import { ConfirmationMethods } from "./types/navigationConfirmation";
-export default function createNavigationConfirmation(): ConfirmationMethods;
