+++ packages/root/types/types/index.d.ts
@@ -2,5 +2,5 @@ export { HistoryOptions, HistoryConstructor, History } from "./hickory";
 export { LocationComponents, PartialLocation, SessionLocation, Hrefable, Key, URLWithState } from "./location";
 export { KeyFns } from "./keyGenerator";
 export { LocationUtilOptions, QueryFunctions, LocationUtils } from "./locationUtils";
-export { NavigationInfo, ConfirmationFunction, ConfirmationMethods } from "./navigationConfirmation";
+export { NavigationInfo, ConfirmationFunction, ConfirmationMethods } from "./confirmation";
 export { Action, NavType, FinishNavigation, CancelNavigation, PendingNavigation, ResponseHandler, Preparer, FinishCancel, NavigateArgs, NavigateHelpers } from "./navigate";
